refactor(rules-service): Split ruleSet and ruleVersion routes
Both features move route/middleware declarations and response Zod
schemas into their own *.routes.ts files; controllers keep only
request-handling logic via createXController(service) factories.
module.ts files now create the shared DocumentedRouter and register
routes into it.

Verified against a correct develop-based baseline (an initial diff run
against a stale already-running process from a different, unmerged
branch briefly showed a false 7th-route mismatch — recaptured the
baseline from this branch's own pre-refactor code, which correctly
shows 6 routes). Route-for-route identical OpenAPI doc before/after,
lint clean, all 29 existing tests pass, live smoke checks pass.
