+++ apps/rules-service/src/rules/ruleSet.routes.ts
@@ -1,0 +1,91 @@
+import { extendZodWithOpenApi } from '@asteasolutions/zod-to-openapi';
+import { z } from 'zod';
+import type { RuleSetService } from './ruleSet.service';
+import { createRuleSetController } from './ruleSet.controller';
+import { createRuleSetSchema } from './dto/create-ruleSet.dto';
+import {
+  requireRoles,
+  trustGatewayIdentity,
+  validateBody,
+  type DocumentedRouter,
+} from '../app.module';
+
+extendZodWithOpenApi(z);
+
+// Request DTO annotated with examples for Swagger UI; validation behavior is
+// unchanged (`.openapi()` only attaches documentation metadata).
+const createRuleSetRequestSchema = createRuleSetSchema.extend({
+  ruleCategory: createRuleSetSchema.shape.ruleCategory.openapi({ example: 'RISK' }),
+  ruleSetName: createRuleSetSchema.shape.ruleSetName.openapi({ example: 'High risk escalation' }),
+  status: createRuleSetSchema.shape.status.openapi({ example: 'DRAFT' }),
+});
+
+const ruleSetSchema = z.object({
+  id: z.string().uuid(),
+  ruleCategory: z
+    .enum(['SCHEDULE', 'RISK', 'ESCALATION', 'INCENTIVE', 'CLOSURE', 'NOTIFICATION'])
+    .openapi({ example: 'RISK' }),
+  ruleSetName: z.string().openapi({ example: 'High risk escalation' }),
+  status: z.enum(['DRAFT', 'ACTIVE', 'RETIRED']).openapi({ example: 'DRAFT' }),
+  createdAt: z.string().datetime(),
+  createdByUserId: z.string().nullable(),
+  updatedAt: z.string().datetime(),
+  updatedByUserId: z.string().nullable(),
+});
+
+const apiErrorSchema = z.object({
+  success: z.literal(false),
+  message: z.string(),
+  errorCode: z.string().openapi({ example: 'VALIDATION_ERROR' }),
+  details: z.record(z.unknown()).optional(),
+});
+
+function envelope<T extends z.ZodTypeAny>(data: T) {
+  return z.object({ success: z.literal(true), message: z.string(), data });
+}
+
+/**
+ * Rule set HTTP routes. Mounted under the global `api/v1` prefix.
+ *
+ * Uses `createDocumentedRouter()` so each route's OpenAPI entry is defined
+ * in the same call as the Express route itself — the request body schema
+ * is inferred from `validateBody` already in the middleware chain, so
+ * `/docs.json` can never drift from what's actually mounted.
+ */
+export function registerRuleSetRoutes(doc: DocumentedRouter, service: RuleSetService) {
+  const controller = createRuleSetController(service);
+
+  doc.get(
+    '/rules',
+    {
+      summary: 'List rule sets',
+      tags: ['Rules'],
+      responses: {
+        200: { description: 'Rule sets', schema: envelope(z.array(ruleSetSchema)) },
+        401: { description: 'Unauthenticated', schema: apiErrorSchema },
+        403: { description: 'Caller role not permitted', schema: apiErrorSchema },
+      },
+    },
+    trustGatewayIdentity,
+    requireRoles('ADMIN'),
+    controller.list,
+  );
+
+  doc.post(
+    '/rules',
+    {
+      summary: 'Create a rule set',
+      tags: ['Rules'],
+      responses: {
+        201: { description: 'Rule set created', schema: envelope(ruleSetSchema) },
+        400: { description: 'Validation error', schema: apiErrorSchema },
+        401: { description: 'Unauthenticated', schema: apiErrorSchema },
+        403: { description: 'Caller role not permitted', schema: apiErrorSchema },
+      },
+    },
+    trustGatewayIdentity,
+    requireRoles('ADMIN'),
+    validateBody(createRuleSetRequestSchema),
+    controller.create,
+  );
+}
